gemini-url-search-tool/requirements.txt
# Core dependencies
streamlit>=1.28.0
requests>=2.31.0
aiohttp>=3.8.0
python-dotenv>=1.0.0

# Data processing
pandas>=2.0.0
beautifulsoup4>=4.12.0
chardet>=5.0.0
plotly>=5.15.0
# sqlite3 is built-in with Python

# AI/ML
google-generativeai>=0.3.0

# Utilities
pydantic>=2.0.0
typing-extensions>=4.5.0

# Development dependencies (optional)
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
flake8>=6.0.0